requirements.txt
Flask==3.0.0
fuzzywuzzy==0.18.0
Jinja2==3.1.2
matplotlib==3.8.2
numpy==1.25.2
python-dotenv==1.0.0
redis==5.0.1
Requests==2.31.0
scikit_learn==1.3.2
SQLAlchemy==2.0.23
vllm==0.2.6
